refactor performance tests into their own file and run them via the command line
