+++ src/lib.rs
@@ -2546,130 +2546,6 @@ mod tests {
     };
     use std::{char, collections::HashMap, str::FromStr, time::Instant};
 
-    fn perf_test_for_db(pool: &AnyPool) {
-        let drop = r#"DROP TABLE IF EXISTS "my_table""#;
-        let create = r#"CREATE TABLE "my_table" (
-                          "row_number" BIGINT,
-                          "prefix" TEXT,
-                          "base" TEXT,
-                          "ontology IRI" TEXT,
-                          "version IRI" TEXT
-                        )"#;
-
-        fn col_to_a1(column_index: isize) -> String {
-            fn divrem(dividend: isize, divisor: isize) -> (isize, isize) {
-                (dividend / divisor, dividend % divisor)
-            }
-            let mut div = column_index;
-            let mut rem;
-            let mut column_id = String::from("");
-            while div > 0 {
-                (div, rem) = divrem(div, 26);
-                if rem == 0 {
-                    rem = 26;
-                    div = -1;
-                }
-                column_id = format!(
-                    "{}{}",
-                    char::from_u32((rem + 64).try_into().unwrap()).unwrap(),
-                    column_id
-                );
-            }
-
-            column_id
-        }
-
-        let mut insert = String::from(r#"INSERT INTO "my_table" VALUES "#);
-        let num_rows = 250000;
-        let num_columns = 5;
-        for i in 1..num_rows {
-            insert.push_str(&format!("({}, ", i));
-            let mut cell_ids = vec![];
-            for j in 1..num_columns {
-                let cell_id = format!("'{}{}'", col_to_a1(j), i);
-                cell_ids.push(cell_id);
-            }
-            insert.push_str(&format!("{})", cell_ids.join(", ")));
-            if i != (num_rows - 1) {
-                insert.push_str(",");
-            }
-        }
-
-        let num_iterations = 5;
-        for i in 1..(num_iterations + 1) {
-            println!(
-                "Running performance test #{} of {} for {:?}",
-                i,
-                num_iterations,
-                pool.any_kind()
-            );
-            for sql in &vec![drop, create, &insert] {
-                let query = sqlx_query(sql);
-                block_on(query.execute(pool)).unwrap();
-            }
-
-            let mut select = Select::new("my_table");
-            select
-                .select(vec!["row_number", "prefix", "base", "\"ontology IRI\"", "\"version IRI\""])
-                .filter(vec![Filter::new("row_number", "lt", json!(num_rows / 2)).unwrap()]);
-
-            // Run the VACUUM command to clear the cache:
-            let query = sqlx_query("VACUUM");
-            block_on(query.execute(pool)).unwrap();
-            // Time the query:
-            let start = Instant::now();
-            let rows = select.fetch_rows(pool, &HashMap::new());
-            println!("Elapsed time for Select::fetch_rows(): {:.2?}", start.elapsed());
-            for row in rows.unwrap() {
-                let _: &str = row.try_get("prefix").unwrap();
-            }
-            println!("Elapsed time after iterating: {:.2?}", start.elapsed());
-
-            // Run the VACUUM command to clear the cache:
-            let query = sqlx_query("VACUUM");
-            block_on(query.execute(pool)).unwrap();
-            // Time the query:
-            let start = Instant::now();
-            let json_rows = select.fetch_rows_as_json(pool, &HashMap::new());
-            println!("Elapsed time for Select::fetch_rows_as_json(): {:.2?}", start.elapsed());
-            for row in json_rows.unwrap() {
-                let _ = row.get("prefix").unwrap();
-            }
-            println!("Elapsed time after iterating: {:.2?}", start.elapsed());
-            println!("----------");
-        }
-    }
-
-    #[test]
-    #[ignore] // Performances test disabled by default.
-    #[serial]
-    /// Runs a performance test of the Select::fetch_rows() and Select::fetch_rows_as_json()
-    /// functions on PostgreSQL. Note that to view the output of the test you must
-    /// use the 'nocapture' option: `cargo test -- --nocapture`
-    fn perf_test_postgresql() {
-        let pg_connection_options =
-            AnyConnectOptions::from_str("postgresql:///valve_postgres").unwrap();
-        let postgresql_pool =
-            block_on(AnyPoolOptions::new().max_connections(5).connect_with(pg_connection_options))
-                .unwrap();
-        perf_test_for_db(&postgresql_pool);
-    }
-
-    #[test]
-    #[ignore] // Performances test disabled by default.
-    #[serial]
-    /// Runs a performance test of the Select::fetch_rows() and Select::fetch_rows_as_json()
-    /// functions on SQLite. Note that to view the output of the test you must
-    /// use the 'nocapture' option: `cargo test -- --nocapture`
-    fn perf_test_sqlite() {
-        let sq_connection_options =
-            AnyConnectOptions::from_str("sqlite://test.db?mode=rwc").unwrap();
-        let sqlite_pool =
-            block_on(AnyPoolOptions::new().max_connections(5).connect_with(sq_connection_options))
-                .unwrap();
-        perf_test_for_db(&sqlite_pool);
-    }
-
     #[test]
     #[serial]
     fn test_real_datatype() {
